Add 'int' cast for getpid() because some Solaris releases return long for getpid().

+++ src/backend/postmaster/pgstat.c
@@ -1505,7 +1505,7 @@ PgstatCollectorMain(int argc, char *argv[])
 	snprintf(pgStat_fname, MAXPGPATH, PGSTAT_STAT_FILENAME, DataDir);
 	/* tmpfname need only be set correctly in this process */
 	snprintf(pgStat_tmpfname, MAXPGPATH, PGSTAT_STAT_TMPFILE,
-			 DataDir, getpid());
+			 DataDir, (int)getpid());
 
 	/*
 	 * Arrange to write the initial status file right away
